Move headers to {Core,Keyboard,Midi} sub-folders
