+++ include/KL/Core/Signal.hpp
