+++ include/KL/Core/detail/Signal.ipp
