+++ src/Core/test/tst_Property.cpp
@@ -11,9 +11,9 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
 
-#include "KL/Property.hpp"
+#include "KL/Core/Property.hpp"
 
-#include "KL/Warnings.hpp"
+#include "KL/Core/Warnings.hpp"
 
 KL_DISABLE_WARNINGS
 #include <catch.hpp>
