+++ src/Keyboard/ComputerKey.cpp
@@ -11,7 +11,7 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
 
-#include "KL/ComputerKey.hpp"
+#include "KL/Keyboard/ComputerKey.hpp"
 
 
 namespace KL
