+++ src/Keyboard/KeyboardLayout.cpp
@@ -11,7 +11,7 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
 
-#include "KL/KeyboardLayout.hpp"
+#include "KL/Keyboard/KeyboardLayout.hpp"
 
 
 namespace KL
