+++ src/Keyboard/platform/default/Keyboard.cpp
@@ -11,7 +11,7 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
 
-#include "KL/Keyboard.hpp"
+#include "KL/Keyboard/Keyboard.hpp"
 
 
 namespace KL
