+++ src/Keyboard/platform/mock/KeyboardPlatformImpl.hpp
@@ -11,8 +11,8 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
 
-#include "KL/Keyboard.hpp"
-#include "KL/Signal.hpp"
+#include "KL/Core/Signal.hpp"
+#include "KL/Keyboard/Keyboard.hpp"
 
 
 namespace KL
