+++ src/Keyboard/test/tst_KeyboardLayout.cpp
@@ -11,9 +11,9 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
 
-#include "KL/KeyboardLayout.hpp"
+#include "KL/Keyboard/KeyboardLayout.hpp"
 
-#include "KL/Warnings.hpp"
+#include "KL/Core/Warnings.hpp"
 
 KL_DISABLE_WARNINGS
 #include <catch.hpp>
