+++ src/Midi/platform/default/MidiOut.cpp
@@ -11,7 +11,7 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
 
-#include "KL/MidiOut.hpp"
+#include "KL/Midi/MidiOut.hpp"
 
 
 namespace KL
