+++ src/TestApp/main.cpp
@@ -11,10 +11,10 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
 
-#include "KL/Keyboard.hpp"
-#include "KL/KeyboardLayout.hpp"
-#include "KL/MidiOut.hpp"
-#include "KL/Warnings.hpp"
+#include "KL/Core/Warnings.hpp"
+#include "KL/Keyboard/Keyboard.hpp"
+#include "KL/Keyboard/KeyboardLayout.hpp"
+#include "KL/Midi/MidiOut.hpp"
 
 KL_DISABLE_WARNINGS
 #include <QtCore/QDebug>
